requirements.txt
altgraph==0.17.4
antiorm==1.2.1
blinker==1.8.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
db==0.1.1
db-sqlite3==0.0.1
dlib==19.24.4
face-recognition==1.3.0
face_recognition_models==0.3.0
face_recognition_util==0.1.2
Flask==3.0.3
future==1.0.0
gevent==24.2.1
greenlet==3.0.3
imageio==2.34.2
itsdangerous==2.2.0
Jinja2==3.1.4
lazy_loader==0.4
MarkupSafe==2.1.5
networkx==3.3
numpy==1.26.0
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pefile==2023.2.7
pillow==10.4.0
ply==3.11
pycparser==2.22
pyinstaller==6.8.0
pyinstaller-hooks-contrib==2024.7
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32-ctypes==0.2.2
scikit-image==0.24.0
scipy==1.14.0
sckit==0.0.5
setuptools==70.2.0
sip==6.8.5
six==1.16.0
tifffile==2024.7.2
tk==0.1.0
tkintertable==1.3.3
toml==0.10.2
tzdata==2024.1
unzip==1.0.0
Werkzeug==3.0.3
wget==3.2
wheel==0.43.0
zope.event==5.0
zope.interface==6.4.post2
